requirements.txt
numpy
pandas
scikit-learn
spacy
gradio
sentence-transformers   # optional (for semantic matching)
dotenv
google.genai